Added dispatchDownloadTask, getMapNameFromDrive
And split the sc2 map reading logic into sc2_map_utils
